Data.Persistent.Collection

Description

A persistent, transactional collection with Queue interface as well as indexed access by key.

Uses default persistence. See Data.TCache.DefaultPersistence

Documentation

type RefQueue a = DBRef (Queue a)

A queue reference

getQRef :: (Typeable a, Serialize a) => String -> RefQueue a

Get the reference to new or existing queue trough its name

pop

Arguments

:: (Typeable a, Serialize a)

=> RefQueue a
Queue name

-> IO a
the returned elems

Read the first element in the queue and delete it (pop)

popSTM :: (Typeable a, Serialize a) => RefQueue a -> STM a

Version in the STM monad

pick

Arguments

:: (Typeable a, Serialize a)

=> RefQueue a
Queue name

-> IO a
the returned elems

flush :: (Typeable a, Serialize a) => RefQueue a -> IO ()

Empty the queue (factually, it is deleted)

flushSTM :: (Typeable a, Serialize a) => RefQueue a -> STM ()

Version in the STM monad

pickAll :: (Typeable a, Serialize a) => RefQueue a -> IO [a]

Return the list of all elements in the queue. The queue remains unchanged

pickAllSTM :: (Typeable a, Serialize a) => RefQueue a -> STM [a]

Version in the STM monad

push :: (Typeable a, Serialize a) => RefQueue a -> a -> IO ()

Push an element in the queue

pushSTM :: (Typeable a, Serialize a) => RefQueue a -> a -> STM ()

Version in the STM monad

pickElem :: (Indexable a, Typeable a, Serialize a) => RefQueue a -> String -> IO (Maybe a)

Return the first element in the queue that has the given key

pickElemSTM :: (Indexable a, Typeable a, Serialize a) => RefQueue a -> String -> STM (Maybe a)

Version in the STM monad

readAll :: (Typeable a, Serialize a) => RefQueue a -> IO [a]

Return the list of all elements in the queue and empty it

readAllSTM :: (Typeable a, Serialize a) => RefQueue a -> STM [a]

A version in the STM monad

deleteElem :: (Indexable a, Typeable a, Serialize a) => RefQueue a -> a -> IO ()

Delete all the elements of the queue that has the key of the parameter passed

deleteElemSTM :: (Typeable a, Serialize a, Indexable a) => RefQueue a -> a -> STM ()

Verison in the STM monad

unreadSTM :: (Typeable a, Serialize a) => RefQueue a -> a -> STM ()

isEmpty :: (Typeable a, Serialize a) => RefQueue a -> IO Bool

Check if the queue is empty

isEmptySTM :: (Typeable a, Serialize a) => RefQueue a -> STM Bool