Data.Persistent.IDynamic

Description

IDynamic is a indexable and serializable version of Dynamic. (See Data.Dynamic). It is used as containers of objects in the cache so any new datatype can be incrementally stored without recompilation. IDimamic provices methods for safe casting, besides serializaton, deserialirezation and retrieval by key.

Documentation

data IDynamic

Constructors

IDyn (IORef IDynType)

data IDynType

Constructors

forall a w r . (Typeable a, Serialize a) => DRight !a

DLeft !(ByteString, (Context, ByteString))

newtype Save

Constructors

Save ByteString

tosave :: IDynamic -> IDynamic

errorfied :: [Char] -> [Char] -> t

dynPrefix :: [Char]

dynPrefixSp :: ByteString

notreified :: ByteString

toIDyn :: (Typeable a, Serialize a) => a -> IDynamic

serializedEqual :: IDynamic -> ByteString -> Bool

fromIDyn :: (Typeable a, Serialize a) => IDynamic -> a

safeFromIDyn :: (Typeable a, Serialize a) => IDynamic -> Either String a

reifyM :: (Typeable a, Serialize a) => IDynamic -> a -> IO a