Data.TCache.DefaultPersistence

Documentation

class Indexable a where

Indexable is an utility class used to derive instances of IResource

Example:

data Person= Person{ pname :: String, cars :: [DBRef Car]} deriving (Show, Read, Typeable)
data Car= Car{owner :: DBRef Person , cname:: String} deriving (Show, Read, Eq, Typeable)

Since Person and Car are instances of Read ans Show, by defining the Indexable instance will implicitly define the IResource instance for file persistence:

instance Indexable Person where  key Person{pname=n} = "Person " ++ n
instance Indexable Car where key Car{cname= n} = "Car " ++ n

Methods

key :: a -> String

defPath

Arguments

:: a

-> String
additional extension for default file paths. IMPORTANT: defPath must depend on the datatype, not the value (must be constant). Default is .tcachedata/

class Serializable a where

Serialize is an alternative to the IResource class for defining persistence in TCache. The deserialization must be as lazy as possible. serialization/deserialization are not performance critical in TCache

Read, Show, instances are implicit instances of Serializable

    serialize  = pack . show
    deserialize= read . unpack

Since write and read to disk of to/from the cache are not be very frequent The performance of serialization is not critical.

Methods

serialize :: a -> ByteString

deserialize :: ByteString -> a

deserialKey :: String -> ByteString -> a

setPersist

Arguments

:: a

-> Maybe Persist
defaultPersist if Nothing

setDefaultPersist :: Persist -> IO ()

Set the default persistence mechanism of all serializable objects that have setPersist= const Nothing. By default it is filePersist

this statement must be the first one before any other TCache call

getDefaultPersist :: Persist

filePersist :: Persist

Implements default default-persistence of objects in files with their keys as filenames

data Persist

a persist mechanism has to implement these three primitives filePersist is the default file persistence

Constructors

Persist
delete

Fields

readByKey :: Key -> IO (Maybe ByteString)

read by key. It must be strict

write :: Key -> ByteString -> IO ()

write. It must be strict

delete :: Key -> IO ()