Data.TCache.Defs

Description

some internal definitions. To use default persistence, use DefaultPersistence instead

Documentation

type AccessTime = Integer

type ModifTime = Integer

data Status a

Constructors

NotRead

DoNotExist

Exist a

data Elem a

Constructors

Elem !a !AccessTime !ModifTime

type TPVar a = TVar (Status (Elem a))

data DBRef a

Constructors

DBRef !String !(TPVar a)

castErr :: (Typeable a1, Typeable a) => a -> a1

class Indexable a where

Indexable is an utility class used to derive instances of IResource

Example:

data Person= Person{ pname :: String, cars :: [DBRef Car]} deriving (Show, Read, Typeable)
data Car= Car{owner :: DBRef Person , cname:: String} deriving (Show, Read, Eq, Typeable)

Since Person and Car are instances of Read ans Show, by defining the Indexable instance will implicitly define the IResource instance for file persistence:

instance Indexable Person where  key Person{pname=n} = "Person " ++ n
instance Indexable Car where key Car{cname= n} = "Car " ++ n

Methods

key :: a -> String

defPath

Arguments

:: a

-> String
additional extension for default file paths. IMPORTANT: defPath must depend on the datatype, not the value (must be constant). Default is TCacheData/

class Serializable a where

Serialize is an alternative to the IResource class for defining persistence in TCache. The deserialization must be as lazy as possible. serialization/deserialization are not performance critical in TCache

Read, Show, instances are implicit instances of Serializable

    serialize  = show
    deserialize= read

Since write and read to disk of to/from the cache are not be very frequent The performance of serialization is not critical.

Methods

serialize :: a -> ByteString

deserialize :: ByteString -> a

setPersist

Arguments

:: a

-> Persist
defaultPersistif not overriden

data Persist

a persist mechanism has to implement these three primitives defaultpersist is the default file persistence

Constructors

Persist
delete

Fields

readByKey :: String -> IO (Maybe ByteString)

read by key. It must be strict

write :: String -> ByteString -> IO ()

write. It must be strict

delete :: String -> IO ()

defaultPersist :: Persist

Implements default persistence of objects in files with their keys as filenames

getPersist :: Serializable a => a -> IO Persist

defaultReadByKey :: String -> IO (Maybe ByteString)

defaultWrite :: String -> ByteString -> IO ()

safeWrite :: [Char] -> ByteString -> IO ()

defaultDelete :: String -> IO ()

defReadResourceByKey :: (Serializable a, Indexable a) => [Char] -> IO (Maybe a)

defWriteResource :: (Serializable a, Indexable a) => a -> IO ()

defDelResource :: (Serializable a, Indexable a) => a -> IO ()

readFileStrict :: FilePath -> IO ByteString

Strict read from file, needed for default file persistence