Data.TCache.IResource

Documentation

class IResource a where

Must be defined for every object to be cached.

Methods

keyResource

Arguments

:: a

-> String
must be defined

readResourceByKey :: String -> IO (Maybe a)

Implements the database access and marshalling of the object. while the database access must be strict, the marshaling must be lazy if, as is often the case, some parts of the object are not really accesed. If the object contains DBRefs, this avoids unnecesary cache lookups. This method is called inside atomically blocks. Since STM transactions retry, readResourceByKey may be called twice in strange situations. So it must be idempotent, not only in the result but also in the effect in the database . However, because it is executed by safeIOToSTM it is guaranteed that the execution is not interrupted.

readResource :: a -> IO (Maybe a)

writeResource :: a -> IO ()

To write into persistent storage. It must be strict. Since STM transactions may retry, writeResource must be idempotent, not only in the result but also in the effect in the database. . However, because it is executed by safeIOToSTM it is guaranteed that the execution is not interrupted. All the new obbects are writeen to the database on synchromization, so writeResource must not autocommit. Commit code must be located in the postcondition. (see setConditions) Since there is no provision for rollback from failure in writing to persistent storage, writeResource must retry until success.

delResource :: a -> IO ()

Delete the resource. It is called syncronously. So it must tocommit

data Resources a b

Resources data definition used by withSTMResources

Constructors

Retry
forces a retry

Resources

Fields

toAdd :: [a]

resources to be inserted back in the cache

toDelete :: [a]

resources to be deleted from the cache and from permanent storage

toReturn :: b

result to be returned

resources :: Resources a ()

Empty resources: resources= Resources [] [] ()