Data.TCache.Memoization

Description

Documentation

writeCached :: (Typeable b, Typeable a, Indexable a, Executable m) => a -> (a -> m b) -> b -> Integer -> STM ()

memoize the result of a computation for a certain time. This is useful for caching costly data such web pages composed on the fly.

time == 0 means infinite

cachedByKey :: (Typeable a, Executable m, MonadIO m) => String -> Int -> m a -> m a

Memoize the result of a computation for a certain time. A string key is used to index the result

The Int parameter is the timeout, in second after the last evaluation, after which the cached value will be discarded and the expression will be evaluated again if demanded . Time == 0 means no timeout

cachedByKeySTM :: (Typeable a, Executable m) => String -> Int -> m a -> STM a

flushCached :: String -> IO ()

cachedp :: (Indexable a, Typeable a, Typeable b) => (a -> b) -> a -> b

a pure version of cached

addrStr :: a -> [Char]

given a string, return a key that can be used in Indexable instances Of non persistent objects, such are cached objects (it changes fron execution to execution) . It uses addrHash

class Executable m where

to execute a monad for the purpose of memoizing its result

Methods

execute :: m a -> a