Data.TCache.Triggers

Documentation

data DBRef a

Constructors

DBRef !String !(TPVar a)

data Elem a

Constructors

Elem !a !AccessTime !ModifTime

data Status a

Constructors

NotRead

DoNotExist

Exist a

addTrigger :: (IResource a, Typeable a) => (DBRef a -> Maybe a -> STM ()) -> IO ()

Add an user defined trigger to the list of triggers Trriggers are called just before an object of the given type is created, modified or deleted. The DBRef to the object and the new value is passed to the trigger. The called trigger function has two parameters: the DBRef being accesed (which still contains the old value), and the new value. If the DBRef is being deleted, the second parameter is Nothing. if the DBRef contains Nothing, then the object is being created

applyTriggers :: (IResource a, Typeable a) => [DBRef a] -> [Maybe a] -> STM ()

internally called when a DBRef is modifieddeletedcreated